extract sharable code into module and superclass
