+++ lib/api/v3/utilities/eager_loading/custom_field_accessor.rb
@@ -30,18 +30,30 @@
 
 module API
   module V3
-    module Projects
-      class ProjectEagerLoadingWrapper < API::V3::Utilities::EagerLoading::EagerLoadingWrapper
-        include API::V3::Utilities::EagerLoading::CustomFieldAccessor
+    module Utilities
+      module EagerLoading
+        module CustomFieldAccessor
+          extend ActiveSupport::Concern
 
-        class << self
-          def wrap(projects)
-            custom_fields = if projects && !projects.empty?
-                              projects.first.available_custom_fields
-                            end
+          included do
+            # Because of the ruby method lookup,
+            # wrapping the work_package here and define the
+            # available_custom_fields methods on the wrapper does not suffice.
+            # We thus extend each work package.
+            def initialize(object)
+              super
+              object.extend(CustomFieldAccessorPatch)
+            end
 
-            super
-              .each { |project| project.available_custom_fields = custom_fields }
+            module CustomFieldAccessorPatch
+              def available_custom_fields
+                @available_custom_fields
+              end
+
+              def available_custom_fields=(fields)
+                @available_custom_fields = fields
+              end
+            end
           end
         end
       end
