+++ lib/api/v3/work_packages/work_package_eager_loading_wrapper.rb
@@ -31,18 +31,9 @@
 module API
   module V3
     module WorkPackages
-      class WorkPackageEagerLoadingWrapper < SimpleDelegator
-        private_class_method :new
-
+      class WorkPackageEagerLoadingWrapper < API::V3::Utilities::EagerLoading::EagerLoadingWrapper
         def wrapped?
           true
-        end
-
-        ##
-        # Workaround against warnings in flatten
-        # delegator does not forward private method #to_ary
-        def to_ary
-          __getobj__.send(:to_ary)
         end
 
         class << self
